feat(#646): use drug search in drug selection
