+++ app/lib/drug_selection/module.dart
@@ -1,9 +1,9 @@
 import '../common/module.dart';
-import 'pages/drugselection.dart';
+import 'pages/drug_selection.dart';
 
 // We need to expose all pages for AutoRouter
 export 'pages/cubit.dart';
-export 'pages/drugselection.dart';
+export 'pages/drug_selection.dart';
 
 const drugSelectionRoutes = AutoRoute(
   path: 'drugselection',
